move OPNewItemField into _new_field_registry module
